backend/knowledge-base/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
sentence-transformers==3.0.1
pinecone-client==4.1.2
PyMuPDF==1.24.10
tqdm==4.66.5
httpx==0.27.2
pydantic==2.8.2
